Move test-utils to lib
