+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/elements/tests/Question.test.js
@@ -2,7 +2,11 @@ import React from "react";
 import { cleanup, render, act, waitFor } from "@testing-library/react";
 import { Question } from "../question/Question";
 import { useTemplateStore } from "@lib/store";
-import { defaultStore as store, Providers, localStorageMock } from "@formbuilder/test-utils";
+import {
+  defaultStore as store,
+  Providers,
+  localStorageMock,
+} from "@lib/utils/form-builder/test-utils";
 import userEvent from "@testing-library/user-event";
 import { LocalizedElementProperties } from "@lib/types/form-builder-types";
